test: refactor duplicate tests for different views
